+++ package.json
@@ -24,7 +24,6 @@
     "react-dom": "^16.13.1",
     "react-gauge-chart": "^0.3.0",
     "react-icons": "^3.10.0",
-    "react-lottie": "^1.2.3",
     "react-pivottable": "^0.10.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
@@ -59,7 +58,6 @@
     "@commitlint/config-conventional": "^8.3.4",
     "@types/react": "^16.14.5",
     "@types/react-gauge-chart": "^0.3.0",
-    "@types/react-lottie": "^1.2.6",
     "@types/styled-components": "^5.1.0",
     "@types/yup": "^0.28.0",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
