+++ src/hooks/useFetch.tsx
@@ -10,7 +10,6 @@ interface Response {
 export function useFetch<Data = any, Error = any>(
   url: string,
   params: object = {},
-  refresh = 120000,
 ): Response {
   const { data, error, mutate } = useSWR<Data, Error>(
     url,
@@ -23,7 +22,7 @@ export function useFetch<Data = any, Error = any>(
       return response.data;
     },
     {
-      refreshInterval: refresh,
+      refreshInterval: 120000,
     },
   );
