--- src/pages/Imports/styles.ts
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  padding-bottom: 16px;
-
-  #pivot {
-    margin: auto;
-    padding: 0 20px 0 20px;
-  }
-`;
