--- src/pages/ProductivityProduction/styles.ts
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  text-align: center;
-
-  h1 {
-    font-size: 96px;
-  }
-
-  strong {
-    background-image: linear-gradient(90deg, #fdd000, #e58a00);
-    background-clip: text;
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-  }
-`;
