+++ src/routes/Route.tsx
@@ -32,21 +32,16 @@ const Route: React.FC<RouterProps> = ({
     <ReactDOMRoute
       {...rest}
       render={({ location }) => {
-        switch (isPrivate) {
-          case !!logged:
-            return <Component />;
-          case false:
-            return <Component />;
-          default:
-            return (
-              <Redirect
-                to={{
-                  pathname: isPrivate ? '/' : `/menu`,
-                  state: { from: location },
-                }}
-              />
-            );
-        }
+        return isPrivate === !!logged ? (
+          <Component />
+        ) : (
+          <Redirect
+            to={{
+              pathname: isPrivate ? '/' : '/menu',
+              state: { from: location },
+            }}
+          />
+        );
       }}
     />
   );
