+++ src/routes/index.tsx
@@ -6,7 +6,6 @@ import Route from './Route';
 
 import SignIn from '../pages/SignIn';
 
-import { useAuth } from '../hooks/auth';
 import SalesMFs from '../pages/SalesMFs';
 import SalesRomp from '../pages/SalesRomp';
 import SalesGer from '../pages/SalesGer';
@@ -17,9 +16,7 @@ import Breakers from '../pages/Breakers';
 import Plates from '../pages/Plates';
 import Menu from '../pages/Menu';
 import Productivity from '../pages/Productivity';
-import ProductivityProduction from '../pages/ProductivityProduction';
-import Revenues from '../pages/Revenues';
-import Imports from '../pages/Imports';
+import { useAuth } from '../hooks/auth';
 
 const Container = styled.div`
   display: flex;
@@ -70,14 +67,8 @@ const Routes: React.FC = () => {
       <Route path="/op-costs" component={OpCosts} isPrivate />
       <Route path="/margin" component={Margin} isPrivate />
       <Route path="/productivity" component={Productivity} isPrivate />
-      <Route
-        path="/productivityproduction"
-        component={ProductivityProduction}
-      />
       <Route path="/breakers" component={Breakers} isPrivate />
       <Route path="/plates" component={Plates} isPrivate />
-      <Route path="/revenues" component={Revenues} isPrivate />
-      <Route path="/imports" component={Imports} />
     </Switch>
   );
 };
